# backend/contracts/rustrag.openapi.yaml
openapi: 3.1.0
info:
  title: RustRAG API
  version: 0.1.0
  description: |
    Runtime API contract for the current RustRAG backend implementation.

    This contract is hand-maintained from the implemented Axum routes in `backend/src/interfaces/http`.
    It aims to reflect the currently available endpoints, request bodies, response shapes,
    auth requirements, and common error envelopes as accurately as practical.
servers:
  - url: http://localhost:8095
    description: Local default
security:
  - bearerAuth: []
...
paths:
...
  /v1/auth/tokens:
    post:
      tags: [auth]
      summary: Mint an API token
      operationId: createToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTokenRequest'
      responses:
        '200':
          description: Newly minted API token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /v1/documents:
    get:
      tags: [documents]
      summary: List documents
      operationId: listDocuments
      security: []
      parameters:
        - $ref: '#/components/parameters/projectIdOptional'
      responses:
        '200':
          description: Document list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentSummary'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [documents]
      summary: Create document metadata record
      operationId: createDocument
      description: Requires `documents:write`, `workspace:admin`, or an `instance_admin` token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDocumentRequest'
      responses:
        '200':
          description: Created document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/content/ingest-text:
    post:
      tags: [content]
      summary: Ingest plain text and create chunks
      operationId: ingestText
      description: Requires `documents:write`, `workspace:admin`, or an `instance_admin` token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestTextRequest'
      responses:
        '200':
          description: Text ingested and chunked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestTextResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/content/search-chunks:
    post:
      tags: [content]
      summary: Lexically search chunks within a project
      operationId: searchChunks
      description: Requires one of `documents:read`, `query:run`, `workspace:admin`, or an `instance_admin` token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchChunksRequest'
      responses:
        '200':
          description: Matching chunk list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChunkResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/content/embed-project:
    post:
      tags: [content]
      summary: Generate embeddings for project chunks
      operationId: embedProjectChunks
      description: |
        Requires `documents:write`, `workspace:admin`, or an `instance_admin` token.

        If `embedding_model_profile_id` is provided, the backend resolves provider/model from that profile;
        otherwise it uses `provider_kind` and `model_name` directly.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmbedProjectChunksRequest'
      responses:
        '200':
          description: Embedding job result summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmbedProjectChunksResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Bearer token minted by `POST /v1/auth/tokens`.
        Note: some read/list endpoints are currently public and do not require auth.
  parameters:
...
    projectIdOptional:
      in: query
      name: project_id
      required: false
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorBody'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorBody'
          examples:
            apiErrorEnvelope:
              value:
                error: unauthorized
        text/plain:
          schema:
            type: string
          examples:
            authExtractorError:
              value: missing authorization header
      headers:
        WWW-Authenticate:
          required: false
          schema:
            type: string
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorBody'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorBody'
        text/plain:
          schema:
            type: string
  schemas:
    ApiErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    CreateTokenRequest:
      type: object
      required: [token_kind, label, scopes]
      properties:
        workspace_id:
          type: [string, 'null']
          format: uuid
        token_kind:
          type: string
          example: workspace_api
        label:
          type: string
          example: frontend-local
        scopes:
          type: array
          items:
            type: string
          example: [documents:read, query:run]
    TokenCreateResponse:
      type: object
      required: [id, token_kind, label, token, scopes]
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: [string, 'null']
          format: uuid
        token_kind:
          type: string
        label:
          type: string
        token:
          type: string
          description: Plaintext token shown only at creation time.
        scopes:
          type: array
          items:
            type: string
...
    DocumentSummary:
      type: object
      required: [id, project_id, external_key]
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        source_id:
          type: [string, 'null']
          format: uuid
        external_key:
          type: string
        title:
          type: [string, 'null']
        mime_type:
          type: [string, 'null']
        checksum:
          type: [string, 'null']
    CreateDocumentRequest:
      type: object
      required: [project_id, external_key]
      properties:
        project_id:
          type: string
          format: uuid
        source_id:
          type: [string, 'null']
          format: uuid
        external_key:
          type: string
        title:
          type: [string, 'null']
        mime_type:
          type: [string, 'null']
        checksum:
          type: [string, 'null']
    IngestTextRequest:
      type: object
      required: [project_id, external_key, text]
      properties:
        project_id:
          type: string
          format: uuid
        source_id:
          type: [string, 'null']
          format: uuid
        external_key:
          type: string
        title:
          type: [string, 'null']
        text:
          type: string
    IngestTextResponse:
      type: object
      required: [document_id, chunk_count]
      properties:
        document_id:
          type: string
          format: uuid
        chunk_count:
          type: integer
          minimum: 0
    SearchChunksRequest:
      type: object
      required: [project_id, query_text]
      properties:
        project_id:
          type: string
          format: uuid
        query_text:
          type: string
        top_k:
          type: [integer, 'null']
          format: int32
    ChunkResult:
      type: object
      required: [id, document_id, ordinal, content]
      properties:
        id:
          type: string
          format: uuid
        document_id:
          type: string
          format: uuid
        ordinal:
          type: integer
          format: int32
        content:
          type: string
    EmbedProjectChunksRequest:
      type: object
      required: [project_id, provider_kind, model_name]
      properties:
        project_id:
          type: string
          format: uuid
        provider_kind:
          type: string
        model_name:
          type: string
        embedding_model_profile_id:
          type: [string, 'null']
          format: uuid
        limit:
          type: [integer, 'null']
          format: int64
    EmbedProjectChunksResponse:
      type: object
      required: [project_id, embedded_chunks, provider_kind, model_name]
      properties:
        project_id:
          type: string
          format: uuid
        embedded_chunks:
          type: integer
          minimum: 0
        provider_kind:
          type: string
        model_name:
          type: string
...